+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### main / head
 
+* Implement `;!` to execute a shell command from the command bar
 * Disabled continuous deployment in GHA. See `.github/workflows/build_artifact.yml`
 for details